fix: redid some bits of the implementation
