+++ src/structures/CommandInteractionOptionResolver.js
@@ -241,15 +241,17 @@ class CommandInteractionOptionResolver {
   }
 
   /**
-   * Gets a focused option.
-   * @param {name} name The name of the option.
+   * Gets the focused option.
    * @param {boolean} [required=false] Whether to throw an error if the option is not found.
-   * @returns {?(string)}
+   * @returns {?(string|number)}
    * The value of the option, or null if not set and not required.
    */
-  getFocused(name, required = false) {
-    const option = this._getTypedOption(name, 'STRING', ['value'], required);
-    return option?.value ?? null;
+  getFocused(required = false) {
+    const focusedOption = this._hoistedOptions.find(option => option.focused);
+    if (required && !focusedOption) {
+      throw new TypeError('AUTOCOMPLETE_INTERACTION_OPTION_NO_FOCUSED_OPTION');
+    }
+    return focusedOption?.value ?? null;
   }
 }
